+++ packages/create-vegas/src/index.ts
@@ -2,6 +2,7 @@
 import child_process from "node:child_process";
 import fs from "node:fs";
 import path from "node:path";
+import util from "node:util";
 
 import * as prompts from "@clack/prompts";
 import { cac } from "cac";
@@ -25,12 +26,11 @@ function runCmd(
   });
 }
 
-const frameworkOptions: {
-  value: string;
-  label?: string | undefined;
-}[] = [
-  { label: "React", value: "template-react" },
-  { label: "Vue", value: "template-vue" },
+const frameworkOptions: { value: string; label: string }[] = [
+  { label: util.styleText("yellow", "Vanilla"), value: "template-vanilla" },
+  { label: util.styleText("green", "Vue"), value: "template-vue" },
+  { label: util.styleText("cyan", "React"), value: "template-react" },
+  { label: util.styleText("magenta", "Preact"), value: "template-preact" },
 ];
 
 function cancelHandler() {
@@ -94,18 +94,6 @@ async function run() {
 
   if (prompts.isCancel(ctx.framework)) {
     cancelHandler();
-  }
-
-  if (ctx.framework === "template-react") {
-    ctx.useOxcStack = (await prompts.confirm({
-      message: "Use Oxc Stack?",
-    })) as boolean;
-
-    if (prompts.isCancel(ctx.useOxcStack)) {
-      cancelHandler();
-    } else if (ctx.useOxcStack) {
-      ctx.framework += "-oxc";
-    }
   }
 
   ctx.npmStartUp = (await prompts.confirm({