+++ packages/create-vegas/template-vanilla/src/server/Code.ts
@@ -2,6 +2,6 @@ export function doGet() {
   const htmlOutput = HtmlService.createHtmlOutputFromFile("index")
   htmlOutput.setFaviconUrl("https://vegasjs.dev/favicon.ico")
   htmlOutput.addMetaTag("viewport", "width=device-width, initial-scale=1.0")
-  htmlOutput.setTitle("Vegas + React + TS")
+  htmlOutput.setTitle("Vegas + TS")
   return htmlOutput
 }