+++ packages/create-vegas/CHANGELOG.md
@@ -1,8 +1,10 @@
-# 0.1.2
+# 0.1.2 (2026-03-22)
 
-chore: improve template.
+chore: we waive the copyright to the template.
+chore: improve react/vue template.
+chore: add vanilla/preact template.
 chore: improve scaffolding prompt.
-chore: add react linter. (react-template-oxc)
+chore: remove template-react-oxc.
 
 # 0.1.1 (2026-03-17)